% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_data}
\alias{download_data}
\title{Download gene information from HGNC, NCBI, and/or Ensembl.}
\usage{
download_data(dir, type = c("all", "hgnc", "ncbi", "ensembl"), force = FALSE)
}
\arguments{
\item{dir}{(a length-1 character vector)
Directory to save the downloaded (default: path/to/extdata/)}

\item{type}{(a length-1 character vector)
Which data to download: "hgnc", "ncbi", "ensembl", or "all" (default: "all")}

\item{force}{(a length-1 logical vector)
Whether to overwrite the existing data (default: FALSE)}
}
\value{
An (invisible) integer code, ‘0’ for success and
non-zero for failure.
}
\description{
Download gene information from HGNC, NCBI, and/or Ensembl.
}
