% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_hgnc}
\alias{download_hgnc}
\title{Download gene information from HGNC.}
\usage{
download_hgnc(dir, force = FALSE)
}
\arguments{
\item{dir}{(a length-1 character vector)
Directory to save the downloaded (default: path/to/extdata/)}

\item{force}{(a length-1 logical vector)
Whether to overwrite the existing data (default: FALSE)}
}
\value{
An (invisible) integer code, ‘0’ for success and
non-zero for failure.
}
\description{
Download gene information from HGNC.
}
