% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/which.genes-package.R
\docType{package}
\name{which.genes-package}
\alias{which.genes}
\alias{which.genes-package}
\title{Which Genes Are They? -- A tool to query gene information.}
\description{
'which.genes' is a tool for querying gene information by
either gene symbols, aliases, or IDs.
}
\details{
Functions:
\itemize{
\item \code{download_data()}
}

Classes:

S3 methods:
}
\author{
\strong{Maintainer}: Chen Liang \email{chen.liang@ghddi.org} (\href{https://orcid.org/0000-0003-0404-1232}{ORCID})

}
\keyword{internal}
